MSITE-115 изменил форматирование цен

+++ lib/EnterModel/Cart/Product.php
@@ -9,7 +9,7 @@ class Product {
     public $ui;
     /** @var int */
     public $quantity;
-    /** @var int */
+    /** @var float */
     public $price;
     /** @var float */
     public $sum;
@@ -27,7 +27,7 @@ class Product {
         if (array_key_exists('id', $data)) $this->id = (string)$data['id'];
         if (array_key_exists('ui', $data)) $this->ui = (string)$data['ui'];
         if (array_key_exists('quantity', $data)) $this->quantity = (int)$data['quantity'];
-        if (array_key_exists('price', $data)) $this->price = (int)$data['price'];
+        if (array_key_exists('price', $data)) $this->price = (float)$data['price'];
         if (array_key_exists('sum', $data)) $this->sum = (float)$data['sum'];
         if (array_key_exists('added', $data)) $this->addedAt = (string)$data['added'];
         if (array_key_exists('sender', $data)) $this->sender = (array)$data['sender'];
